Run the test only if /dev/tty exists.
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@33857 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ test/ruby/test_io.rb
@@ -2126,6 +2126,7 @@ End
   def test_ioctl_linux2
     return if /linux/ !~ RUBY_PLATFORM
     return if /^i?86|^x86_64/ !~ RUBY_PLATFORM
+    return if File.exist?('/dev/tty')
 
     File.open('/dev/tty') { |f|
       tiocgwinsz=0x5413
